+++ src/me/xnuminousx/spirits/Abilities/API/DarkAbility.java
@@ -1,4 +1,4 @@
-package me.xnuminousx.spirits.ability.API;
+package me.xnuminousx.spirits.Abilities.API;
 
 import org.bukkit.entity.Player;
 
@@ -9,9 +9,9 @@ import com.projectkorra.projectkorra.ability.SubAbility;
 
 import me.xnuminousx.spirits.Elements.SpiritElement;
 
-public abstract class LightAbility extends ElementalAbility implements SubAbility {
+public abstract class DarkAbility extends ElementalAbility implements SubAbility {
 
-	public LightAbility(Player player) {
+	public DarkAbility(Player player) {
 		super(player);
 	}
 	
@@ -22,6 +22,6 @@ public abstract class LightAbility extends ElementalAbility implements SubAbilit
 	
 	@Override
 	public Element getElement() {
-		return SpiritElement.LIGHT_SPIRIT;
+		return SpiritElement.DARK_SPIRIT;
 	}
 }
